scripts/backtest/requirements.txt
# 並發回測框架 V2 依賴

# 核心依賴 (V1 也需要)
pandas>=1.3.0
openpyxl>=3.0.0
openai>=1.0.0
psycopg2-binary>=2.9.0
requests>=2.26.0

# V2 新增依賴
aiohttp>=3.8.0           # 異步 HTTP 客戶端
aiodns>=3.0.0            # 異步 DNS 解析 (提升 aiohttp 性能)
tqdm>=4.62.0             # 進度條顯示
tenacity>=8.0.0          # 智能重試機制

# 可選依賴
# python-dotenv>=0.19.0  # 環境變數管理
# prometheus-client      # Prometheus 監控
